[refactor] #186 Add extension function to create dialog
