+++ app/src/main/java/com/afume/afume_android/ui/home/MoreNewPerfumeActivity.kt
@@ -24,7 +24,7 @@ class MoreNewPerfumeActivity : AppCompatActivity() {
     }
 
     private fun initNewList(){
-        newAdapter = MoreNewListAdapter(supportFragmentManager) { idx -> homeViewModel.postPerfumeLike(2, idx)}
+        newAdapter = MoreNewListAdapter(this,supportFragmentManager) { idx -> homeViewModel.postPerfumeLike(2, idx)}
         binding.rvHomeMoreNew.adapter = newAdapter
     }
